+++ src/clojure/nrepl/middleware.clj
@@ -122,7 +122,7 @@
                                  ;; only conj'ing m here to support direct reference to
                                  ;; middleware dependencies in :expects and :requires,
                                  ;; e.g. interruptable-eval's dep on
-                                 ;; nrepl.middleware.pr-values/pr-values
+                                 ;; nrepl.middleware.print/wrap-print
                                  (update-in [:handles] (comp set #(conj % m) keys))
                                  (assoc :implemented-by m))))]
       (set (for [m middlewares]