+++ CHANGELOG.md
@@ -13,8 +13,16 @@
 
 * [#16](https://github.com/nrepl/nrepl/issues/16): Use a single session thread per evaluation.
 * [#107](https://github.com/nrepl/nrepl/issues/107): Stop reading and evaluating code on first read error.
-* Support single-arity printer functions.
 * Restore the `nrepl.bencode` namespace.
+* [#117](https://github.com/nrepl/nrepl/issues/117): Replace
+  `nrepl.middleware.pr-values` with `nrepl.middleware.print`.
+  * New dynamic vars in `nrepl.middleware.print` for configuring the print
+    middleware at the REPL. See the Misc page in the Usage section of the
+    documentation for more information.
+  * The new middleware provides behaviour that is backwards-compatible with the
+    old one. Existing middleware descriptors whose `:requires` set contains
+    `#'pr-values` should instead use `#'wrap-print`. See the Middleware page in
+    the Design section of the documentation for more information.
 
 ### 0.5.3 (2018-12-12)